agent/requirements.txt
pyirsdk>=1.5.0
websockets>=13.0
